Adds tests for main page, cleans up DISCUSSION.md.
