+++ src/app/components/AdvocatesTable.tsx
@@ -1,16 +1,6 @@
 "use client";
 
-// src/app/types.ts (optional, for sharing types)
-export type Advocate = {
-  id?: string | number;
-  firstName: string;
-  lastName: string;
-  city: string;
-  degree: string;
-  specialties: string[];
-  yearsOfExperience: number;
-  phoneNumber: string;
-};
+import { Advocate } from "@/app/types";
 
 type AdvocatesTableProps = {
   advocates: Advocate[];
