+++ src/app/types.ts
@@ -1,0 +1,10 @@
+export type Advocate = {
+  id?: string | number;
+  firstName: string;
+  lastName: string;
+  city: string;
+  degree: string;
+  specialties: string[];
+  yearsOfExperience: number;
+  phoneNumber: string;
+};
